==> notebooks/2_feature_engineering.ipynb <==
# Ingeniería de Características

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MinMaxScaler
import pickle

## 1. Cargamos dataset

df = pd.read_csv('../data/raw/diabetes.csv')
df.head(10)

## 2. Dividimos en train y test

X = df.drop('Outcome', axis=1)  
y = df['Outcome']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

df_train = pd.concat([X_train, y_train], axis=1)
df_train

### Guardamos los datasets de test despúes del split

df_test = pd.concat([X_test, y_test], axis=1)
df_test.to_csv('../data/processed/df_test.csv', index=False)

## 3. Eliminamos variables con muchos faltantes

df_train.drop(['SkinThickness', 'Insulin'], axis=1, inplace=True)
df_train.head()

## 4. Ingeniería de características

### 4.1 Imputación de variables

proporcion_ceros = (df_train == 0).mean()
proporcion_ceros

cols_imputacion = ["Glucose", "BloodPressure", "BMI"]

feature_eng_configs = {}

for col in cols_imputacion:
    media = int(df_train[col].mean())
    imputed_key = f"{col}_imputed_value"  
    feature_eng_configs[imputed_key] = media
    df_train[col] = df_train[col].replace(0, media).astype(int)

### Guardamos valores imputados como artefacto

with open("../artifacts/feature_eng_configs.pkl", "wb") as f:
    pickle.dump(feature_eng_configs, f)

### 4.2 Escalado de variables

mm_scaler = MinMaxScaler()
mm_scaler.fit(df_train)

df_scaled = pd.DataFrame(mm_scaler.transform(df_train), columns=df_train.columns)

### Guardamos el Scaler como artefacto

with open("../artifacts/mm_scaler.pkl", "wb") as f:
    pickle.dump(mm_scaler, f)

## 5. Guardamos dataset procesado

df_scaled.to_csv('../data/processed/features_for_model.csv', index=False)
